+++ lib/api/builder.ts
@@ -1,6 +1,6 @@
 'use client';
 
-import { AxiosError, AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
+import { AxiosInstance, AxiosResponse, AxiosRequestConfig } from 'axios';
 
 import { useEffect, useState } from 'react';
 import { IS_DEBUG } from '../directive';
@@ -14,67 +14,22 @@ const buildUrlWithParams = (baseUrl: string, req: Record<string, string | number
   return params !== '' ? `${baseUrl}?${params}` : baseUrl;
 };
 
-export type ApiResponse<T> =
-  | {
-      status: 'SUCCESS' | 'ERROR';
-      statusCode: number;
-      data: T;
-      message: string;
-      version: string;
-    }
-  | {
-      status: 'FAILURE';
-      statusCode: number;
-      data: 'NO_RES' | 'NO_REQ' | 'FAIL_RES';
-      message: string;
-      version: string;
-    };
+type Builder = <Req, Res>(
+  instance: AxiosInstance,
+  method: 'POST' | 'GET',
+  path: string,
+  config?: AxiosRequestConfig,
+) => ApiCall<Req, Res>;
 
-function failWith(data: 'NO_RES' | 'NO_REQ' | 'FAIL_RES', message: string, statusCode: number): ApiResponse<unknown> {
-  return {
-    status: 'FAILURE',
-    statusCode,
-    data,
-    message,
-    version: 'NO VERSION, This response was generated by Academ Fronted',
-  };
-}
-
-/**
- * AxiosResponse Promise를 받아서 항상 resolve된 버전의 Promise로 바꿔줌
- */
-function resolvify<Res>(promise: Promise<AxiosResponse<ApiResponse<Res>>>) {
+function mould<Res>(promise: Promise<AxiosResponse<ApiResponse<Res>>>) {
   return promise
     .then((res) => {
       const data: ApiResponse<Res> = res.data;
       data.statusCode = res.status;
       return Promise.resolve(res.data as ApiResponse<Res>);
     })
-    .catch((error: AxiosError) => {
-      if (error.response) {
-        return Promise.resolve(
-          failWith('FAIL_RES', '실패한 응답을 받았습니다.', error.response.status) as ApiResponse<Res>,
-        );
-      } else if (error.request) {
-        return Promise.resolve(
-          failWith(
-            'NO_RES',
-            '응답을 수신하지 못했습니다. 장치가 네트워크에 연결되어 있지 않거나 서버가 오프라인입니다.',
-            -2,
-          ) as ApiResponse<Res>,
-        );
-      } else {
-        return Promise.resolve(
-          failWith(
-            'NO_REQ',
-            '요청을 전송하지 못했습니다. 장치가 네트워크에 연결 되지 않았나요?',
-            -3,
-          ) as ApiResponse<Res>,
-        );
-      }
-    })
-    .catch(() => {
-      return Promise.resolve(failWith('NO_REQ', '알 수 없는 오류', -4) as ApiResponse<Res>);
+    .catch((error) => {
+      return Promise.reject(error);
     });
 }
 
@@ -98,7 +53,7 @@ export const build: Builder = <Req, Res>(
         ? instance.post(path, req, config)
         : instance.get(buildUrlWithParams(path, req as Record<string, string | number>), config);
 
-    const firstResult = await resolvify<Res>(firstTry);
+    const firstResult = await mould<Res>(firstTry);
 
     if (IS_DEBUG) {
       const show = firstResult.status === 'SUCCESS' ? console.log : console.error;
@@ -108,24 +63,6 @@ export const build: Builder = <Req, Res>(
     return firstResult;
   };
 };
-
-export type ApiCall<Req, Res> = (req: Req) => Promise<ApiResponse<Res>>;
-export type Builder = <Req, Res>(
-  instance: AxiosInstance,
-  method: 'POST' | 'GET',
-  path: string,
-  config?: AxiosRequestConfig,
-) => ApiCall<Req, Res>;
-
-export type ApiState<T> =
-  | {
-      loading: true;
-      response: unknown;
-    }
-  | {
-      loading: false;
-      response: ApiResponse<T>;
-    };
 
 export function useApi<Req, Res>(apiCall: ApiCall<Req, Res>, req: Req) {
   const [loading, setLoading] = useState<boolean>(true);
