+++ lib/api/calls/membership.ts
@@ -1,4 +1,4 @@
 import { build } from '@/lib/api/builder';
-import withTokenInstance from './instances.ts/withTokenInstance';
+import withTokenInstance from '../instances/withTokenInstance';
 
 export const apiBuyAcess = build<{ item: string }, null>(withTokenInstance, 'POST', '/api/mypage/buy-access-authority');
